+++ examples/onoff-plug-app/silabs/BUILD.gn
@@ -116,8 +116,6 @@ silabs_executable("onoff-plug-app") {
   sources = [
     "${examples_common_plat_dir}/main.cpp",
     "src/AppTask.cpp",
-    "src/DataModelCallbacks.cpp",
-    "src/OnOffPlugManager.cpp",
   ]
 
   deps = [
