+++ examples/onoff-plug-app/silabs/src/AppTask.cpp
@@ -22,6 +22,8 @@
 #include "AppEvent.h"
 #include "LEDWidget.h"
 
+#include <app-common/zap-generated/ids/Attributes.h>
+#include <app-common/zap-generated/ids/Clusters.h>
 #include <app/clusters/on-off-server/on-off-server.h>
 #include <app/server/Server.h>
 #include <app/util/attribute-storage.h>
@@ -33,10 +35,15 @@
 #include <setup_payload/SetupPayload.h>
 
 #include <lib/support/CodeUtils.h>
+#include <lib/support/TypeTraits.h>
 
 #include <platform/CHIPDeviceLayer.h>
 
 #include <platform/silabs/platformAbstraction/SilabsPlatform.h>
+
+#ifdef DIC_ENABLE
+#include "dic.h"
+#endif // DIC_ENABLE
 
 #ifdef SL_CATALOG_SIMPLE_LED_LED1_PRESENT
 #define ONOFF_LED 1
@@ -48,17 +55,25 @@
 #define APP_ONOFF_BUTTON 1
 
 using namespace chip;
+using namespace chip::TLV;
+using namespace chip::app::Clusters;
+using namespace chip::app::Clusters::OnOff;
 using namespace ::chip::DeviceLayer;
 using namespace ::chip::DeviceLayer::Silabs;
 
 namespace {
 
 LEDWidget sOnOffLED;
+TimerHandle_t sPlugTimer;
+
+OnOffEffect gEffect = {
+    chip::EndpointId{ 1 },
+    AppTask::OnTriggerOffWithEffect,
+    EffectIdentifierEnum::kDelayedAllOff,
+    to_underlying(DelayedAllOffEffectVariantEnum::kDelayedOffFastFade),
+};
 
 } // namespace
-
-using namespace chip::TLV;
-using namespace ::chip::DeviceLayer;
 
 AppTask AppTask::sAppTask;
 
@@ -67,21 +82,35 @@ CHIP_ERROR AppTask::AppInit()
     CHIP_ERROR err = CHIP_NO_ERROR;
     chip::DeviceLayer::Silabs::GetPlatform().SetButtonsCb(AppTask::ButtonEventHandler);
 
-    err = PlugMgr().Init();
-    if (err != CHIP_NO_ERROR)
+    sPlugTimer = xTimerCreate("plugTmr",
+                              pdMS_TO_TICKS(1),
+                              false,
+                              (void *) this,
+                              TimerEventHandler);
+
+    if (sPlugTimer == NULL)
     {
-        SILABS_LOG("PlugMgr().Init() failed");
-        appError(err);
+        SILABS_LOG("sPlugTimer timer create failed");
+        return APP_ERROR_CREATE_TIMER_FAILED;
     }
 
-    PlugMgr().SetCallbacks(ActionCallback);
+    bool currentLedState;
+    chip::DeviceLayer::PlatformMgr().LockChipStack();
+    OnOffServer::Instance().getOnOffValue(1, &currentLedState);
+    chip::DeviceLayer::PlatformMgr().UnlockChipStack();
+
+    mIsOn                = currentLedState;
+    mAutoTurnOff         = false;
+    mAutoTurnOffDuration = 0;
+    mOffEffectArmed      = false;
+    mAutoTurnOffTimerArmed = false;
 
     sOnOffLED.Init(ONOFF_LED);
-    sOnOffLED.Set(PlugMgr().IsPlugOn());
+    sOnOffLED.Set(IsPlugOn());
 
 // Update the LCD with the Stored value. Show QR Code if not provisioned
 #ifdef DISPLAY_ENABLED
-    GetLCD().WriteDemoUI(PlugMgr().IsPlugOn());
+    GetLCD().WriteDemoUI(IsPlugOn());
 #ifdef QR_CODE_ENABLED
 #ifdef SL_WIFI
     if (!ConnectivityMgr().IsWiFiStationProvisioned())
@@ -134,12 +163,12 @@ void AppTask::AppTaskMain(void * pvParameter)
 void AppTask::OnOffActionEventHandler(AppEvent * aEvent)
 {
     bool initiated = false;
-    OnOffPlugManager::Action_t action;
+    Action_t action;
     CHIP_ERROR err = CHIP_NO_ERROR;
 
     if (aEvent->Type == AppEvent::kEventType_Button)
     {
-        action = (PlugMgr().IsPlugOn()) ? OnOffPlugManager::OFF_ACTION : OnOffPlugManager::ON_ACTION;
+        action = (sAppTask.IsPlugOn()) ? OFF_ACTION : ON_ACTION;
     }
     else
     {
@@ -148,7 +177,7 @@ void AppTask::OnOffActionEventHandler(AppEvent * aEvent)
 
     if (err == CHIP_NO_ERROR)
     {
-        initiated = PlugMgr().InitiateAction(aEvent->Type, action);
+        initiated = sAppTask.InitiateAction(aEvent->Type, action);
 
         if (!initiated)
         {
@@ -175,10 +204,9 @@ void AppTask::ButtonEventHandler(uint8_t button, uint8_t btnAction)
     }
 }
 
-void AppTask::ActionCallback(OnOffPlugManager::Action_t aAction, int32_t aActor)
+void AppTask::ActionCallback(Action_t aAction, int32_t aActor)
 {
-    // Action initiated, update the light led
-    bool lightOn = aAction == OnOffPlugManager::ON_ACTION;
+    bool lightOn = aAction == ON_ACTION;
     SILABS_LOG("Turning light %s", (lightOn) ? "On" : "Off")
 
     sOnOffLED.Set(lightOn);
@@ -187,11 +215,11 @@ void AppTask::ActionCallback(OnOffPlugManager::Action_t aAction, int32_t aActor)
     sAppTask.GetLCD().WriteDemoUI(lightOn);
 #endif
 
-    if (aAction == OnOffPlugManager::ON_ACTION)
+    if (aAction == ON_ACTION)
     {
         SILABS_LOG("Outlet ON")
     }
-    else if (aAction == OnOffPlugManager::OFF_ACTION)
+    else if (aAction == OFF_ACTION)
     {
         SILABS_LOG("Outlet OFF")
     }
@@ -204,13 +232,209 @@ void AppTask::ActionCallback(OnOffPlugManager::Action_t aAction, int32_t aActor)
 
 void AppTask::UpdateClusterState(intptr_t context)
 {
-    uint8_t newValue = PlugMgr().IsPlugOn();
+    uint8_t newValue = sAppTask.IsPlugOn();
 
-    // write the new on/off value
     Protocols::InteractionModel::Status status = OnOffServer::Instance().setOnOffValue(1, newValue, false);
 
     if (status != Protocols::InteractionModel::Status::Success)
     {
         SILABS_LOG("ERR: updating on/off %x", to_underlying(status));
+    }
+}
+
+bool AppTask::IsPlugOn() const
+{
+    return mIsOn;
+}
+
+void AppTask::EnableAutoTurnOff(bool aOn)
+{
+    mAutoTurnOff = aOn;
+}
+
+void AppTask::SetAutoTurnOffDuration(uint32_t aDurationInSecs)
+{
+    mAutoTurnOffDuration = aDurationInSecs;
+}
+
+bool AppTask::InitiateAction(int32_t aActor, Action_t aAction)
+{
+    bool action_initiated = false;
+
+    if (((mIsOn == false) || mOffEffectArmed) && aAction == ON_ACTION)
+    {
+        action_initiated = true;
+        if (mOffEffectArmed)
+        {
+            CancelTimer();
+            mOffEffectArmed = false;
+        }
+    }
+    else if (mIsOn && aAction == OFF_ACTION && mOffEffectArmed == false)
+    {
+        action_initiated = true;
+        if (mAutoTurnOffTimerArmed)
+        {
+            mAutoTurnOffTimerArmed = false;
+            CancelTimer();
+        }
+    }
+
+    if (action_initiated)
+    {
+        mIsOn = (aAction == ON_ACTION);
+
+        ActionCallback(aAction, aActor);
+
+        if (mAutoTurnOff && mIsOn)
+        {
+            StartTimer(mAutoTurnOffDuration * 1000);
+            mAutoTurnOffTimerArmed = true;
+            SILABS_LOG("Auto Turn off enabled. Will be triggered in %u seconds", mAutoTurnOffDuration);
+        }
+    }
+
+    return action_initiated;
+}
+
+void AppTask::StartTimer(uint32_t aTimeoutMs)
+{
+    if (xTimerIsTimerActive(sPlugTimer))
+    {
+        SILABS_LOG("app timer already started!");
+        CancelTimer();
+    }
+
+    if (xTimerChangePeriod(sPlugTimer, pdMS_TO_TICKS(aTimeoutMs), 100) != pdPASS)
+    {
+        SILABS_LOG("sPlugTimer timer start() failed");
+        appError(APP_ERROR_START_TIMER_FAILED);
+    }
+}
+
+void AppTask::CancelTimer(void)
+{
+    if (xTimerStop(sPlugTimer, pdMS_TO_TICKS(0)) == pdFAIL)
+    {
+        SILABS_LOG("sPlugTimer stop() failed");
+        appError(APP_ERROR_STOP_TIMER_FAILED);
+    }
+}
+
+void AppTask::TimerEventHandler(TimerHandle_t xTimer)
+{
+    AppTask * app = static_cast<AppTask *>(pvTimerGetTimerID(xTimer));
+
+    AppEvent event;
+    event.Type               = AppEvent::kEventType_Timer;
+    event.TimerEvent.Context = app;
+    if (app->mAutoTurnOffTimerArmed)
+    {
+        event.Handler = AutoTurnOffTimerEventHandler;
+    }
+    else if (app->mOffEffectArmed)
+    {
+        event.Handler = OffEffectTimerEventHandler;
+    }
+    AppTask::GetAppTask().PostEvent(&event);
+}
+
+void AppTask::AutoTurnOffTimerEventHandler(AppEvent * aEvent)
+{
+    AppTask * app = static_cast<AppTask *>(aEvent->TimerEvent.Context);
+    int32_t actor = AppEvent::kEventType_Timer;
+
+    if (!app->mAutoTurnOffTimerArmed)
+    {
+        return;
+    }
+
+    app->mAutoTurnOffTimerArmed = false;
+
+    SILABS_LOG("Auto Turn Off has been triggered!");
+
+    app->InitiateAction(actor, OFF_ACTION);
+}
+
+void AppTask::OffEffectTimerEventHandler(AppEvent * aEvent)
+{
+    AppTask * app = static_cast<AppTask *>(aEvent->TimerEvent.Context);
+    int32_t actor = AppEvent::kEventType_Timer;
+
+    if (!app->mOffEffectArmed)
+    {
+        return;
+    }
+
+    app->mOffEffectArmed = false;
+
+    SILABS_LOG("OffEffect completed");
+
+    app->InitiateAction(actor, OFF_ACTION);
+}
+
+void AppTask::OnTriggerOffWithEffect(OnOffEffect * effect)
+{
+    auto effectId              = effect->mEffectIdentifier;
+    auto effectVariant         = effect->mEffectVariant;
+    uint32_t offEffectDuration = 0;
+
+    if (effectId == EffectIdentifierEnum::kDelayedAllOff)
+    {
+        auto typedEffectVariant = static_cast<DelayedAllOffEffectVariantEnum>(effectVariant);
+        if (typedEffectVariant == DelayedAllOffEffectVariantEnum::kDelayedOffFastFade)
+        {
+            offEffectDuration = 800;
+            ChipLogProgress(Zcl, "DelayedAllOffEffectVariantEnum::kDelayedOffFastFade");
+        }
+        else if (typedEffectVariant == DelayedAllOffEffectVariantEnum::kNoFade)
+        {
+            offEffectDuration = 800;
+            ChipLogProgress(Zcl, "DelayedAllOffEffectVariantEnum::kNoFade");
+        }
+        else if (typedEffectVariant == DelayedAllOffEffectVariantEnum::kDelayedOffSlowFade)
+        {
+            offEffectDuration = 12800;
+            ChipLogProgress(Zcl, "DelayedAllOffEffectVariantEnum::kDelayedOffSlowFade");
+        }
+    }
+    else if (effectId == EffectIdentifierEnum::kDyingLight)
+    {
+        auto typedEffectVariant = static_cast<DyingLightEffectVariantEnum>(effectVariant);
+        if (typedEffectVariant == DyingLightEffectVariantEnum::kDyingLightFadeOff)
+        {
+            offEffectDuration = 1500;
+            ChipLogProgress(Zcl, "DyingLightEffectVariantEnum::kDyingLightFadeOff");
+        }
+    }
+
+    sAppTask.mOffEffectArmed = true;
+    sAppTask.StartTimer(offEffectDuration);
+}
+
+void MatterPostAttributeChangeCallback(const chip::app::ConcreteAttributePath & attributePath, uint8_t type, uint16_t size,
+                                       uint8_t * value)
+{
+    ClusterId clusterId     = attributePath.mClusterId;
+    AttributeId attributeId = attributePath.mAttributeId;
+    ChipLogProgress(Zcl, "Cluster callback: " ChipLogFormatMEI, ChipLogValueMEI(clusterId));
+
+    if (clusterId == OnOff::Id && attributeId == OnOff::Attributes::OnOff::Id)
+    {
+#ifdef DIC_ENABLE
+        dic_sendmsg("light/state", (const char *) (value ? (*value ? "on" : "off") : "invalid"));
+#endif // DIC_ENABLE
+
+        sAppTask.InitiateAction(AppEvent::kEventType_Plug, *value ? AppTask::ON_ACTION : AppTask::OFF_ACTION);
+    }
+    else if (clusterId == LevelControl::Id)
+    {
+        ChipLogProgress(Zcl, "Level Control attribute ID: " ChipLogFormatMEI " Type: %u Value: %u, length %u",
+                        ChipLogValueMEI(attributeId), type, *value, size);
+    }
+    else if (clusterId == Identify::Id)
+    {
+        ChipLogProgress(Zcl, "Identify attribute ID: " ChipLogFormatMEI " Type: %u Value: %u, length %u",
+                        ChipLogValueMEI(attributeId), type, *value, size);
     }
 }
